requirements.txt
newsapi-python
yfinance
beautifulsoup4
pandas
requests
